Cache last found number of power supplies
* cbatticon.c:
(changed_power_supplies): Add old_num_ps, old_total_ps, num_ps, and
total_ps to cache and track the number of power supplies.

Cache the last found number of power supplies so that changes in the
number of power supplies are detected correctly. This maintains the
old behavior of also being true if the number of total power supplies
stays the same, but the one that the current cbatticon process is
meant to get changes.

https://github.com/valr/cbatticon/issues/30

Issue #30 should be fixed by this change

Signed-off-by: Jeff Mickey <j@codemac.net>